fix: scope DNS signing-store trust checks to Linux
