+++ docs/src/routes/configuration/+page.md
@@ -141,7 +141,7 @@ TCP 80 in the firewall; binding privileged ports may require OS permissions.
 
 Serves the relay base domain from an authoritative DNS server embedded in the relay process, so no DNS provider API credentials are required. It is the default provider when `ACME_DNS_PROVIDER` is unset. Delegate the base domain once at the parent zone (`NS portal.example.com -> ns.portal.example.com` with glue `A` pointing at the relay public IP) and open `53/tcp` + `53/udp`. Containers running without root need `CAP_NET_BIND_SERVICE` to bind the default port. A answers for the apex and every covered name are synthesized from the relay public IPv4; ACME DNS-01 TXT and tunnel ECH HTTPS records are served directly. DNSSEC signing is always enabled; ENS TXT automation remains opt-in with `ENS_GASLESS_ENABLED=true`.
 
-The relay automatically generates a single ECDSA P-256 CSK (DNSSEC algorithm 13) in `IDENTITY_PATH/dnssec-csk.json`. Preserve and back up this file with the identity volume across restarts, container replacement, and migration: deleting or replacing it changes the DNSKEY and breaks validation against an existing parent DS. On Unix the key is created with mode `0600` (new directories `0700`); permissive existing keys, symlinks, malformed keys, and keys for another zone fail startup rather than triggering automatic replacement. On Windows a protected DACL limits the key file to the running account and SYSTEM, including temporary key files from creation. On Windows, only directories created by the embedded provider receive private directory ACLs. Existing directory and sibling-file ACLs are left unchanged: ownership and write/control access must be trusted to the running account, SYSTEM, or built-in Administrators; unrelated read/list/traverse access is allowed. Unsafe or unsupported Windows directory policies fail startup rather than being silently rewritten; correct those policies explicitly before restarting. On Linux and other supported Unix systems, the configured signing directory and existing key must be owned by the effective service UID or root. Directories with group/other write bits fail startup, including writable Linux POSIX access-ACL mask bits; safe read/traverse sharing remains allowed. This deliberately rejects a writable ACL mask even when its current entries are more restrictive. Existing directories, keys, and sibling files are inspected without changing their ownership or permissions. Correct unsafe storage explicitly before restarting; the same checks run for new-key creation and existing-key loads. Use trusted ancestor directories on every platform. Operators and cleanup jobs must not rename, remove, or replace signing-store directories during initialization; failed-creation rollback coordinates cooperating relay initializers, not external namespace replacement.
+The relay automatically generates a single ECDSA P-256 CSK (DNSSEC algorithm 13) in `IDENTITY_PATH/dnssec-csk.json`. Preserve and back up this file with the identity volume across restarts, container replacement, and migration: deleting or replacing it changes the DNSKEY and breaks validation against an existing parent DS. On Unix the key is created with mode `0600` (new directories `0700`); permissive existing keys, symlinks, malformed keys, and keys for another zone fail startup rather than triggering automatic replacement. On Windows a protected DACL limits the key file to the running account and SYSTEM, including temporary key files from creation. On Windows, only directories created by the embedded provider receive private directory ACLs. Existing directory and sibling-file ACLs are left unchanged: ownership and write/control access must be trusted to the running account, SYSTEM, or built-in Administrators; unrelated read/list/traverse access is allowed. Unsafe or unsupported Windows directory policies fail startup rather than being silently rewritten; correct those policies explicitly before restarting. On Linux, the configured signing directory and existing key must be owned by the effective service UID or root. Directories with group/other write bits fail startup, including writable Linux POSIX access-ACL mask bits; safe read/traverse sharing remains allowed. This deliberately rejects a writable ACL mask even when its current entries are more restrictive. Existing directories, keys, and sibling files are inspected without changing their ownership or permissions. Correct unsafe storage explicitly before restarting; Linux checks run before temporary-key creation and on existing-key loads. macOS and other non-Linux, non-Windows targets retain the existing mode-based private-key check, but do not validate directory ownership or extended ACLs. In particular, macOS extended ACLs can grant access despite restrictive mode bits: operators must independently secure both the directory and key ownership/ACLs. The Linux ACL-mask policy is not a macOS security guarantee. Use trusted ancestor directories on every platform. Operators and cleanup jobs must not rename, remove, or replace signing-store directories during initialization; failed-creation rollback coordinates cooperating relay initializers, not external namespace replacement.
 
 The signing store must permit its required file and directory durability operations, including when an existing key is loaded. A visible key name may still belong to a concurrent creator whose publication has not finished flushing; read-only or search-only storage that cannot establish this durability barrier is not supported. Flush failures fail startup without replacing the key.
